bundle adj must be here to avoid circular import
